Move agent custom instalation code to tests. It's the only place where it's used
